security: implement audit remediation across backend + admin panel
Harden auth/session handling (refresh rotation + grace window, idempotent session), tighten CORS/CSRF, add rate limiting + login lockout, and make key order/inventory flows transaction-safe and atomic. Improve admin UI safety (sanitized errors, pagination, destructive-action confirms) and add log/serialization redaction to prevent sensitive data leakage.

+++ src/controllers/review.controller.js
@@ -4,7 +4,7 @@ import { Product } from '../models/product.model.js';
 
 const createSchema = z.object({
   productId: z.string().trim().min(1, 'productId is required').max(120),
-  rating: z.number().min(1).max(5),
+  rating: z.number().int().min(1).max(5),
   comment: z.string().trim().max(4000).optional(),
 });
